+++ src/Bankly.php
@@ -13,6 +13,7 @@ use WeDevBr\Bankly\Support\Contracts\DocumentInterface;
 use WeDevBr\Bankly\Types\Billet\DepositBillet;
 use WeDevBr\Bankly\Types\Pix\PixEntries;
 use WeDevBr\Bankly\Types\Card\Card;
+use WeDevBr\Bankly\Contracts\Pix\PixCashoutInterface;
 
 /**
  * Class Bankly
@@ -457,6 +458,16 @@ class Bankly
     public function deletePixAddressingKeyValue(string $addressingKeyValue)
     {
         return $this->delete("/pix/entries/$addressingKeyValue");
+    }
+
+    /**
+     * @param PixCashoutInterface $pixCashout
+     * @param string $correlationId
+     * @return array|mixed
+     */
+    public function pixCashout(PixCashoutInterface $pixCashout, string $correlationId)
+    {
+        return $this->post('/pix/cash-out', $pixCashout->toArray(), $correlationId, true);
     }
 
     /**