+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 All notable changes to `bankly-laravel` will be documented in this file
-## 1.8.0 - 2021-05-06
+
+## 1.8.0 - 2021-05-14
 - Added deposit billet management
+- Added PIX cashout Manual
+- Added PIX cashout StaticQrCode
+- Fix features tests
 
 ## 1.7.0 - 2021-04-27
 - Refactor Virtual Card